requirements.txt
google-cloud-aiplatform>=1.60.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sentence-transformers>=2.6.0
httpx>=0.27.0
pydantic>=2.6.0
tenacity>=8.2.3
jinja2>=3.1.3
python-dateutil>=2.9.0
requests>=2.31.0
flask>=3.0.0
numpy>=1.26.0
